Joel dev

+++ src/main/java/pe/edu/cibertec/RC/Repository/CartDao.java
@@ -3,5 +3,5 @@ package pe.edu.cibertec.RC.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.cibertec.RC.Model.Carrito;
 
-public interface CartDao extends JpaRepository<Carrito, Long> {
+public interface CartDao extends JpaRepository<Carrito, Integer> {
 }
